samples/grounding/file-search-tool/requirements.txt
google-genai==1.49.0